+++ HISTORY.md
@@ -3,7 +3,7 @@
 ## Version 0.6.0
 
 - Release date: Dec 07, 2023
-- added-custom-log-func-and-refactor-tx-error-handling-v0.6.0
+- added-custom-log-func-and-refactor-accessInfoType-tx-error-handling-v0.6.0
 
 
 ## Version 0.5.1